gameStruct was rewrite, all animations are documented and working(ZOMBIE, STANDARD AND SPECIAL), linear interpolation using leap is done, but has some issues that need to be fixed, renderEMD need to be rewrite

+++ gameMath.h
@@ -26,7 +26,7 @@ public:
   	bool collisionDetect(unsigned int colType, signed int x_1, signed int z_1, signed int w, signed int h, signed int x, signed int z);
   	bool mapSwitch(signed int x, signed int z, signed int x1, signed int z1, signed int x2, signed int z2, signed int x3, signed int z3, signed int x4, signed int z4);
 
-  	int interpolation(int n1, int n2, int p);
+  	float interpolation(float n1, float n2, float p);
 private:
